feat: add drag & drop import functionality for external note files in sidebar
- Implemented a new module for handling drag and drop of external note files (.md, .html, .txt, .json, .zip) onto the notes sidebar.
- Added support for importing files into specific folders or at the root level based on drop location.
- Included user feedback through overlays and notifications during the import process.
- Ensured compatibility with read-only workspaces and provided error handling for unsupported file types.
