+++ src/i18n/ru.json
@@ -1411,7 +1411,20 @@
       "complete_backup": "Перетащите ZIP-файл резервной копии сюда",
       "database": "Перетащите SQL-файл сюда",
       "notes": "Перетащите ZIP-файл с заметками сюда",
-      "attachments": "Перетащите ZIP-файл с вложениями сюда"
+      "attachments": "Перетащите ZIP-файл с вложениями сюда",
+      "sidebar": {
+        "drop_here": "Перетащите файлы заметок, чтобы импортировать их",
+        "drop_into_folder": "Отпустите, чтобы импортировать в \"{{folder}}\"",
+        "folder_not_allowed": "Невозможно импортировать в \"{{folder}}\"",
+        "importing": "Импорт заметок...",
+        "import_done": "Заметки импортированы.",
+        "import_failed": "Ошибка импорта.",
+        "no_supported_files": "Не было перетащено ни одного поддерживаемого файла. Разрешённые типы: {{allowed}}",
+        "skipped_files": "Пропущено неподдерживаемых файлов: {{count}}."
+      },
+      "errors": {
+        "system_folder": "Заметки нельзя импортировать в папку \"{{folder}}\"."
+      }
     },
     "modals": {
       "import_confirm": {
